# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EASY GAS API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base path"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "usersList",
        "summary": "List users (employee directory)",
        "description": "Requires permission: users.view. Branch-scoped for non-admin roles. The CURRENT viewer is excluded from their own directory (before COUNT and pagination), so the total and pages never include the caller — every other user (including other administrators) is listed. Deterministic order (created_at desc, id desc). Use GET /users/me for the caller's own profile.",
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RoleCode"
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/UserStatus"
            }
          },
          {
            "$ref": "#/components/parameters/PageQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Classic paginated user list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "usersCreate",
        "summary": "Create user",
        "description": "Requires permission: users.create.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "phone",
                  "region",
                  "role",
                  "password"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Demo"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Demo"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+998900000000"
                  },
                  "region": {
                    "type": "string",
                    "example": "Toshkent"
                  },
                  "role": {
                    "$ref": "#/components/schemas/RoleCode"
                  },
                  "branchId": {
                    "type": [
                      "integer",
                      "null"
                    ]
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "usersGetOwnProfile",
        "summary": "Get own profile",
        "description": "Authenticated only — self-scoped, so NO users.view permission is required (every employee may view their own profile). Returns the caller's own UserDetail (branch name, status, created/last-login). It never accepts a target id and cannot read another account.",
        "responses": {
          "200": {
            "description": "Own profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/IdPath"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "usersGet",
        "summary": "Get user",
        "description": "Requires permission: users.view. Out-of-scope / nonexistent ids return 404 (never 403) so ids cannot be probed across branch boundaries.",
        "responses": {
          "200": {
            "description": "User detail.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "usersUpdate",
        "summary": "Update user",
        "description": "Requires permission: users.update.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "role": {
                    "$ref": "#/components/schemas/RoleCode"
                  },
                  "branchId": {
                    "type": [
                      "integer",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/users/{id}/block": {
      "parameters": [
        {
          "$ref": "#/components/parameters/IdPath"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "usersBlock",
        "summary": "Block user",
        "description": "Requires permission: users.block.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "User blocked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/users/{id}/unblock": {
      "parameters": [
        {
          "$ref": "#/components/parameters/IdPath"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "usersUnblock",
        "summary": "Unblock user",
        "description": "Requires permission: users.block.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "User unblocked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/users/{id}/reset-password": {
      "parameters": [
        {
          "$ref": "#/components/parameters/IdPath"
        }
      ],
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "usersResetPassword",
        "summary": "Issue a temporary password (manual admin recovery)",
        "description": "ADMIN only — requires permission: users.reset_password. Manual employee password recovery: after verifying the employee OUT OF BAND (an admin sends a support request to @EasygasGarantbot; the bot is NOT an automated auth channel), the admin issues a one-time temporary password for a specific employee. The admin re-confirms their OWN current password and gives a mandatory reason. The server generates a cryptographically-random temporary password, atomically sets it (hashed) with must_change_password + a configurable expiry, revokes ALL the employee's sessions, and invalidates any prior recovery credentials. A repeated reset invalidates the previous temporary password. The temporary password is returned ONCE in the response (Cache-Control: no-store) and is NEVER logged, cached, put in a URL, or recorded in the audit chain — the audit records only actor/target/reason/timestamp. An admin cannot reset their OWN password here. Blocked / inactive-branch employees stay blocked.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "reason"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "The admin's own current password (re-auth)."
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temporary password issued (returned once).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/UserDetail"
                    },
                    "temporaryPassword": {
                      "type": "string",
                      "description": "One-time secret — display to the admin once; never persisted client-side."
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "eg_session",
        "description": "HttpOnly session cookie set by POST /auth/login. State-changing requests must also send the x-csrf-token header (see the CSRF model in the API description)."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "PageQuery": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 20
        }
      },
...
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "CsrfTokenHeader": {
        "name": "x-csrf-token",
        "in": "header",
        "required": true,
        "description": "CSRF token = HMAC-SHA256(APP_KEY, \"csrf:\" + raw session cookie). Required on all state-changing (POST/PATCH/DELETE) requests that carry the session cookie. Invalid/missing token or bad origin returns 403 (CSRF_TOKEN / CSRF_ORIGIN).",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Not authenticated or invalid credentials. Stable codes: UNAUTHORIZED.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted, or CSRF check failed. Stable codes: FORBIDDEN, CSRF_TOKEN, CSRF_ORIGIN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found (or hidden by branch scope). Stable codes: NOT_FOUND.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "State conflict. Stable codes: CONFLICT, RISK_POLICY_NOT_APPROVED (fail-closed when no risk matrix is active), SUMMARY_STALE (signable summary changed).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request failed validation. Stable codes: VALIDATION_ERROR, GPS_INVALID (details[].code one of NAN_OR_INFINITE, OUT_OF_RANGE, NULL_ISLAND, LOW_ACCURACY, BAD_TIMESTAMP, FUTURE_TIMESTAMP, STALE), OUT_OF_RANGE (checklist measurement out of range).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded. Stable codes: TOO_MANY_REQUESTS. See the Retry-After / RateLimit-* headers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable business error code."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "RoleCode": {
        "type": "string",
        "enum": [
          "USTA",
          "MASTER",
          "RAHBAR",
          "SIFAT",
          "ADMIN"
        ]
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "BLOCKED"
        ]
      },
...
      "User": {
        "type": "object",
        "description": "Authenticated user with the server-issued permission list.",
        "required": [
          "id",
          "firstName",
          "lastName",
          "phone",
          "region",
          "role",
          "status",
          "permissions"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "branchId": {
            "type": [
              "integer",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/RoleCode"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserDetail": {
        "type": "object",
        "required": [
          "id",
          "firstName",
          "lastName",
          "phone",
          "region",
          "role",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/RoleCode"
          },
          "branchId": {
            "type": [
              "integer",
              "null"
            ]
          },
          "branchName": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "lastLoginAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ClassicListMeta": {
        "type": "object",
        "description": "Classic CRUD list envelope metadata: { total, page, limit }.",
        "required": [
          "total",
          "page",
          "limit"
        ],
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
...
      "UsersListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClassicListMeta"
          },
          {
            "type": "object",
            "required": [
              "users"
            ],
            "properties": {
              "users": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}